+++ src/mass_driver/cli.py
@@ -1,7 +1,10 @@
 """Command line entrypoint for mass-driver"""
 import argparse
 import sys
+from pathlib import Path
 from typing import Optional
+
+from mass_driver.drivers import Counter
 
 
 def parse_arguments(arguments: list[str]) -> argparse.Namespace:
@@ -10,7 +13,16 @@ def parse_arguments(arguments: list[str]) -> argparse.Namespace:
         "mass-driver",
         description="Send bulk repo change requests",
     )
-    parser.add_argument("foo", help="Some parameter")
+    parser.add_argument("repo_path", help="Repository to Patch, as Git Url")
+
+    detect_group = parser.add_mutually_exclusive_group()
+    detect_group.add_argument(
+        "--patch", action="store_true", help="Actually do the patching"
+    )
+    detect_group.add_argument(
+        "--detect", action="store_true", help="Just detect, no patching"
+    )
+    parser.set_defaults(patch=False, detect=True)
     return parser.parse_args(arguments)
 
 
@@ -19,9 +31,16 @@ def cli(arguments: Optional[list[str]] = None):
     if arguments is None:
         arguments = sys.argv[1:]
     args = parse_arguments(arguments)
-    main(args.foo)
+    repo_path = Path(args.repo_path)
+    main(repo_path, args.patch)
 
 
-def main(foo):
+def main(repo_path: Path, do_patch: bool):
     """Run the program's main command"""
-    print(f"Foo is: {foo}")
+    driver = Counter(counter_file=Path("counter"), target_count=1)
+    if do_patch:
+        print(f"Patching '{repo_path}'...")
+        driver.patch(repo_path)
+    else:
+        needs_patch = driver.detect(repo_path)
+        print(f"Detecting '{repo_path}': {needs_patch=}")