+++ src/mass_driver/drivers/__init__.py
@@ -1,0 +1,4 @@
+"""All the Patch Drivers we've implemented"""
+from mass_driver.drivers.counter import Counter
+
+__all__ = ["Counter"]