+++ src/mass_driver/patch_driver.py
@@ -1,0 +1,25 @@
+"""The core classes for creating Patches across Repositories"""
+
+from pathlib import Path
+
+GitUrl = str
+"""A git-cloneable URL, HTTPS, GIT protocol or even local path"""
+
+
+class PatchDriver:
+    """Base class for creating patches over repositories"""
+
+    patch_parameters: dict
+    """Parameters given to the driver"""
+
+    def detect(self, repo_path: Path) -> bool:
+        """
+        Scan the given (cloned) Git Repository for patch-ablity.
+
+        Return True if the repo is "non-compliant", needing a patch.
+        """
+        raise NotImplementedError("PatchDriver base class can't detect, use derived")
+
+    def patch(self, repo_path: Path):
+        """Mutate the given (cloned) repo to Patch it out"""
+        raise NotImplementedError("PatchDriver base class can't patch, use derived")