+++ CHANGELOG.md
@@ -6,6 +6,9 @@ The project uses semantic versioning (see [semver](https://semver.org)).
 
 ## [Unreleased]
 
+### Added
+- Basic PatchDriver system, to detect/patch local repos
+
 
 ## v0.1.0 - 2022-11-10
 ### Added